src/ai-quest-api/requirements.txt
anthropic
fastapi
mangum
python-dotenv
